pyproject.toml
[tool.poetry]
name = "quarto-aap-mottakere"
version = "0.1.0"
packages = [{ include = "hjelpefunksjoner" }]
authors = ["team Spenn i Nav IT"]
description = "Datafortelling i Quarto for offentlig statistikk på mottakere av arbeidsavklaringspenger (AAP)"
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.0.1"
pandas = "^2.2.2"
plotly = "^5.22.0"
sqlfluff = "^3.1.0"
oracledb = "^2.2.1"
tabulate = "^0.9.0"
dvh-tools = "^0.7.3"
db-dtypes = "^1.2.0"
jupyterlab = "^4.2.4"
google-cloud-storage = "^2.17.0"
google-cloud-secret-manager = "^2.20.1"
google-api-python-client = "^2.137.0"
ipywidgets = "^8.1.3"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
